Add NewRequest function
Not all http clients that support the Doer interface can be expected to handle authentication themselves - it’s sometimes more convent to pass a request object off to a third party client.

This change exposes a NewRequest method so, if the user chooses, they’re able to manage the client outside of this package.